refactor(capabilities): split 1335-LOC god module into 5 sub-modules
Pure relocation — no behavior changes. metrics_config/capabilities.rs
(1,335 LOC) is decomposed into capabilities/ with diagnostics/semantic/
runtime_health/entrypoint sub-modules. Tests are co-located in
entrypoint.rs as #[cfg(test)] mod tests so super::* reuses module-level
imports without manual rebinding.

Visibility cost: DiagnosticsGuidance.status field + for_file /
guidance_payload / unavailable_payload methods raised from private to
pub(super) to span the new sub-module boundary. mod.rs preserves the
external surface (get_capabilities, collect_runtime_health_snapshot,
SemanticSearchStatus) so cross-crate callers and metrics_config/tests.rs
require zero changes.

Verification:
- cargo check -p codelens-mcp --features http,semantic — PASS
- cargo test -p codelens-mcp --bin codelens-mcp metrics_config — 15/15 PASS
- cargo fmt --all — applied
