+++ crates/codelens-mcp/src/tools/session/metrics_config/capabilities/mod.rs
@@ -1,0 +1,9 @@
+mod diagnostics;
+mod entrypoint;
+mod runtime_health;
+mod semantic;
+
+pub use entrypoint::get_capabilities;
+pub(crate) use runtime_health::collect_runtime_health_snapshot;
+#[cfg(test)]
+pub(crate) use semantic::SemanticSearchStatus;
